# Add an iterator to `Distribution`