This tries to solve the same problem as https://github.com/rust-lang-nursery/rand/pull/275, having a nice way to sample from distributions using an iterator. See also https://github.com/rust-lang-nursery/rand/issues/58.

I have added a `sample_iter` method to `Distribution` for now. But I don't mind if it is a method on `Rng` instead, or on both...

With benchmarks similar to those in https://github.com/rust-lang-nursery/rand/pull/275, the results are all over the place. In all situations it is slightly slower than `iter::repeat`, but usually not by more than a couple of percent. The exception is `Uniform`, where we use that values generated by the RNG directly:
```
test gen_1k_fill                   ... bench:       1,129 ns/iter (+/- 1) = 906 MB/s
test gen_1k_gen_iter               ... bench:         729 ns/iter (+/- 1) = 1404 MB/s
test gen_1k_iter_repeat            ... bench:         372 ns/iter (+/- 0) = 2752 MB/s
test gen_1k_sample_iter            ... bench:         626 ns/iter (+/- 1) = 1635 MB/s <- new
```

Example of using `Distribution::sample_iter` instead of `Rng::gen_iter`:
```rust
use distributions::{Alphanumeric, Range, Uniform};
let rng = thread_rng();

// Vec of 16 x f32:
// old
let v: Vec<f32> = iter::repeat(()).map(|()| rng.gen()).take(16).collect();
// new
let v: Vec<f32> = Uniform.sample_iter(&mut rng).take(16).collect();

// String
// old
let s: String = rng.gen_ascii_chars().take(10).collect();
// new
let s: String = Alphanumeric.sample_iter(&mut rng).take(7).collect();

// Combined values
// old
println!("{:?}", rng.gen_iter::<(f64, bool)>().take(5)
                    .collect::<Vec<(f64, bool)>>());
// new
println!("{:?}", Uniform.sample_iter(&mut rng).take(5)
                        .collect::<Vec<(f64, bool)>>());

// Dice-rolling:
// old
let die_range = Range::new_inclusive(1, 6);
let mut roll_die = iter::repeat(()).map(|()| rng.sample(die_range));
while roll_die.next().unwrap() != 6 {
    println!("Not a 6; rolling again!");
}
// new
let die_range = Range::new_inclusive(1, 6);
let mut roll_die = die_range.sample_iter(&mut rng);
while roll_die.next().unwrap() != 6 {
    println!("Not a 6; rolling again!");
}
```